# Conflicts:
#	market-api/src/main/java/com/maxcar/statistics/controller/RankingController.java

+++ market-statistics/market-statistics-interface/src/main/java/com/maxcar/statistics/model/entity/InventoryInvoiceMonthEntity.java
@@ -17,7 +17,7 @@ public class InventoryInvoiceMonthEntity implements Serializable {
 
     private Double stockPrice;
 
-    private Integer stockDayAvg;
+    private Double stockDayAvg;
 
     private Integer outLibraryCount;
 
@@ -103,11 +103,11 @@ public class InventoryInvoiceMonthEntity implements Serializable {
         this.stockPrice = stockPrice;
     }
 
-    public Integer getStockDayAvg() {
+    public Double getStockDayAvg() {
         return stockDayAvg;
     }
 
-    public void setStockDayAvg(Integer stockDayAvg) {
+    public void setStockDayAvg(Double stockDayAvg) {
         this.stockDayAvg = stockDayAvg;
     }
